Fixed fmi1_import_get_value_referece_list on 64 bit. See ticket #2292. This could be detected earlier just by looking into the warnings. Therefore went through all the warnings on 64-bit and cleaned up most of them.

+++ Test/jm_vector_test.c
@@ -53,7 +53,7 @@ int main() {
 		if(jm_vector_get_size(int)(v) != VINIT_SIZE+i+1) log_error("Vector size %d is not as expected %d\n", jm_vector_get_size(int)(v), VINIT_SIZE+i+1);
     }
 	{
-		int index = jm_vector_find_index(int)(v, &k,jm_compare_int);
+		size_t index = jm_vector_find_index(int)(v, &k,jm_compare_int);
 		k = TESTVAL;
 		if( index != 2) log_error("Index of '%d' should be '2' but got %d\n", TESTVAL, k );
 	}
